spes: props: Make SurfaceFlinger latch all buffers unsignaled
* According to [1], LatchUnsignaledConfig in A13 defaults to AutoSingleLayer, so latching unsignaled buffers is permitted when a single
layer is updated in a frame, and the update includes just a buffer update with no sync transactions or geometry changes. However, this
may still be suboptimal for legacy devices with highly limited hardware capabilities, resulting in inconsistent ui performance, since
the time it takes for all the other buffers to become signaled can be pretty high. Hence in order to maintain consistent ui performance,
change the LatchUnsignaledConfig to Always by setting debug.sf.latch_unsignaled prop to true.

[1] - https://source.android.com/docs/core/graphics/unsignaled-buffer-latch

Signed-off-by: angelomds42 <angelomds42@gmail.com>